+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added the ability to track a metric other than "loss" for validation in `TorchTrainStep` ("torch::train").
+
 ## [v0.1.2](https://github.com/allenai/tango/releases/tag/v0.1.2) - 2021-10-13
 
 ### Added